Pod mutator only acts on qsts pods
Also clean up test helper files, make test more robust by waiting for
all pods and avoid docker.io even more.

[#176379221](https://www.pivotaltracker.com/story/show/176379221)
